% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_regression.R
\name{ridgereg}
\alias{ridgereg}
\title{Ridge Regression}
\usage{
ridgereg(formula, data, lambda = 1)
}
\arguments{
\item{formula}{A formula object}

\item{data}{A data frame}

\item{lambda}{Regularization parameter}
}
\value{
An object of class "ridgereg"
}
\description{
Ridge Regression
}
