% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr_ridge.R
\name{ridgereg_qr}
\alias{ridgereg_qr}
\title{Ridge Regression via QR}
\usage{
ridgereg_qr(formula, data, lambda = 1)
}
\arguments{
\item{formula}{A formula object}

\item{data}{A data frame}

\item{lambda}{Regularization parameter}
}
\value{
An object of class "ridgereg"
}
\description{
Ridge Regression via QR
}
