% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\name{ridgereg}
\alias{ridgereg}
\title{Ridge Regression}
\usage{
ridgereg(formula, data, lambda = 1)
}
\arguments{
\item{formula}{A formula object}

\item{data}{A data frame}

\item{lambda}{Regularization parameter}
}
\value{
An object of class "ridgereg"
}
\description{
Ridge Regression
}
\examples{
data(iris)
model <- ridgereg(Petal.Length ~ Sepal.Length + Sepal.Width, data = iris, lambda = 0.1)
print(model)
predict(model, iris)
coef(model)
}
