% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg_qr.R
\name{ridgereg_qr}
\alias{ridgereg_qr}
\title{Ridge Regression using QR decomposition}
\usage{
ridgereg_qr(formula, data, lambda = 1)
}
\arguments{
\item{formula}{A formula object}

\item{data}{A data frame}

\item{lambda}{Regularization parameter}
}
\value{
An object of class "ridgereg_qr"
}
\description{
Ridge Regression using QR decomposition
}
\examples{
data(iris)
model <- ridgereg_qr(Petal.Length ~ Sepal.Length + Sepal.Width, data = iris, lambda = 0.1)
print(model)
predict(model, iris)
coef(model)
}
